Move logic from sectionProgressRedux to a sectionProgressLoader
