+++ apps/src/templates/sectionProgress/sectionProgressConstants.js
@@ -27,3 +27,6 @@ export const ViewType = {
   DETAIL: 'detail', // levels
   STANDARDS: 'standards'
 };
+
+export const tooltipIdForStudent = studentId =>
+  `tooltipIdForStudent${studentId}`;
